chore: migrate version announcement modal
